requirements.txt
# data-generator
pyyaml
psycopg2-binary
faker
python-dotenv
requests

# Consumer
kafka-python
boto3
fastparquet
pandas